residence_controller_edit_changes

+++ app/Http/Controllers/Admin/ResidenceController.php
@@ -105,8 +105,10 @@ class ResidenceController extends BaseController
         $req = $request->all();
         $req['amenities'] = json_encode($request->amenities);
 
-        if(isset($req['grid_image'])) {
+        if(isset($req['grid_image']) && file_exists(public_path('uploads/'.$model->grid_image))) {
             unlink(public_path() . '/uploads/' . $model->grid_image);
+            $req['grid_image'] = $this->fileUpload($request->grid_image, public_path('uploads/'))[0];
+        }else {
             $req['grid_image'] = $this->fileUpload($request->grid_image, public_path('uploads/'))[0];
         }
